New include system added to core files
