+++ src/console_io.c
@@ -1,5 +1,9 @@
 #include "console_io.h"
 #include "AI/detect.h"
+#include "core/groups_s.h"
+#include "core/hand_s.h"
+#include "core/histogram_s.h"
+#include "core/riichi_engine_s.h"
 #include "debug.h"
 #include <stdio.h>
 #include <wchar.h>
