+++ src/main.c
@@ -1,5 +1,6 @@
 #define _POSIX_C_SOURCE 199309L
 #include "core/riichi_engine.h"
+#include "core/riichi_engine_s.h"
 #include <locale.h>
 #include <stdio.h>
 #include <stdlib.h>
